rename LogoutButton.tsx
